Homework4 Spring Framework

+++ src/main/java/org/habitApp/services/HabitService.java
@@ -1,31 +1,53 @@
 package org.habitApp.services;
 
 import org.habitApp.annotations.Loggable;
-import org.habitApp.dto.habitDto.HabitReportDto;
-import org.habitApp.models.Habit;
+import org.habitApp.domain.dto.habitDto.HabitReportDto;
+import org.habitApp.domain.entities.HabitEntity;
+import org.habitApp.domain.entities.UserEntity;
+import org.habitApp.exceptions.HabitNotFoundException;
+import org.habitApp.exceptions.UnauthorizedAccessException;
+import org.habitApp.exceptions.HabitAlreadyCompletedException;
 import org.habitApp.models.Period;
-import org.habitApp.models.User;
 import org.habitApp.repositories.HabitComletionHistoryRepository;
 import org.habitApp.repositories.HabitRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import java.sql.SQLException;
 import java.time.LocalDate;
 import java.time.temporal.ChronoUnit;
 import java.util.List;
 import java.util.Objects;
-import java.util.UUID;
 
 /**
  * Сервис для управления привычками (CRUD)
  * Позволяет создавать, редактировать, удалять и просматривать привычки пользователя.
  */
+@Loggable
+@Service
 public class HabitService {
+    @Autowired
     private final HabitRepository habitRepository;
+
+    @Autowired
     private final HabitComletionHistoryRepository habitComletionHistoryRepository;
 
     public HabitService(HabitRepository habitRepository, HabitComletionHistoryRepository habitComletionHistoryRepository) {
         this.habitRepository = habitRepository;
         this.habitComletionHistoryRepository = habitComletionHistoryRepository;
+    }
+
+    public HabitEntity getHabitById(long habitId, UserEntity currentUser) throws SQLException, HabitNotFoundException, UnauthorizedAccessException {
+        HabitEntity habit = habitRepository.getHabitById(habitId);
+        if (habit == null) {
+            throw new HabitNotFoundException("Привычка с ID " + habitId + " не найдена.");
+        }
+
+        // Проверка принадлежности привычки текущему пользователю
+        if (habit.getUserId() != (currentUser.getId())) {
+            throw new UnauthorizedAccessException("Привычка не принадлежит текущему пользователю.");
+        }
+        return habit;
     }
 
     /**
@@ -35,10 +57,9 @@ public class HabitService {
      * @param name        название привычки
      * @param description описание привычки
      * @param frequency   частота выполнения привычки
-     * @throws SQLException ошибка работы с БД
      */
-    public void createHabit(User user, String name, String description, Period frequency) throws SQLException {
-        Habit habit = Habit.CreateHabit(name, description, frequency.getPeriodName(), LocalDate.now(), user.getId());
+    public void createHabit(UserEntity user, String name, String description, Period frequency) throws SQLException {
+        HabitEntity habit = new HabitEntity(name, description, frequency.getPeriodName(), LocalDate.now(), user.getId());
         habitRepository.createHabit(habit);
         System.out.println("Привычка \"" + name + "\" создана для пользователя " + user.getEmail() + ".");
     }
@@ -50,10 +71,15 @@ public class HabitService {
      * @param newName        новое название
      * @param newDescription новое описание
      * @param newFrequency   новая частота выполнения
-     * @throws SQLException ошибка работы с БД
      */
-    public void updateHabit(UUID habitId, String newName, String newDescription, Period newFrequency) throws SQLException {
-        Habit habit = new Habit(habitId, newName, newDescription, newFrequency.toString(), null, null);
+    public void updateHabit(long habitId, String newName, String newDescription, Period newFrequency) throws SQLException {
+        HabitEntity habit = habitRepository.getHabitById(habitId);
+        if (habit == null) {
+            throw new HabitNotFoundException("Habit not found.");
+        }
+        habit.setName(newName);
+        habit.setDescription(newDescription);
+        habit.setFrequency(newFrequency.toString());
 
         habitRepository.updateHabit(habitId, habit);
         System.out.println("Привычка обновлена: " + habit.getName());
@@ -63,15 +89,19 @@ public class HabitService {
      * Удаление привычки
      *
      * @param habitId привычка для удаления
-     * @throws SQLException ошибка работы с БД
      */
-    public void deleteHabit(UUID habitId, User currentUser) throws SQLException {
-        UUID userId = habitRepository.getHabitById(habitId).getUserId();
+    public void deleteHabit(long habitId, UserEntity currentUser) throws SQLException {
+        HabitEntity habit = habitRepository.getHabitById(habitId);
+        if (habit == null) {
+            throw new HabitNotFoundException("Habit not found.");
+        }
+
+        long userId = habit.getUserId();
         if (userId == currentUser.getId()) {
             habitRepository.deleteHabit(habitId);
-            System.out.println("Привычка \"" + habitId + "\" была удалена.");
+            System.out.println("Привычка \"" + habit.getName() + "\" была удалена.");
         } else {
-            System.out.println("Привычка не принадлежит текущему пользователю");
+            throw new UnauthorizedAccessException("Habit does not belong to the current user.");
         }
     }
 
@@ -80,9 +110,8 @@ public class HabitService {
      *
      * @param user текущий пользователь
      * @return список привычек пользователя
-     * @throws SQLException ошибка работы с БД
      */
-    public List<Habit> getAllHabits(User user) throws SQLException {
+    public List<HabitEntity> getAllHabits(UserEntity user) throws SQLException {
         return habitRepository.getHabitsByUser(user);
     }
 
@@ -91,12 +120,10 @@ public class HabitService {
      *
      * @param currentUser текущий пользователь
      * @return список всех привычек
-     * @throws SQLException           ошибка работы с БД
-     * @throws IllegalAccessException доступ для админа
      */
-    public List<Habit> getAllHabitsAdmin(User currentUser) throws SQLException, IllegalAccessException {
-        if (!currentUser.isAdmin()) {
-            throw new IllegalAccessException("User is not admin");
+    public List<HabitEntity> getAllHabitsAdmin(UserEntity currentUser) throws SQLException {
+        if (!currentUser.isFlagAdmin()) {
+            throw new UnauthorizedAccessException("User is not admin.");
         }
         return habitRepository.getAllHabits();
     }
@@ -104,18 +131,16 @@ public class HabitService {
     /**
      * Отметить привычку как выполненную
      *
-     * @param habit привычка, которую нужно отметить сегодня
+     * @param habitId привычка, которую нужно отметить сегодня
      */
-    public void markHabitAsCompleted(Habit habit) throws SQLException {
-        List<LocalDate> completionHistory = habitComletionHistoryRepository.getCompletionHistoryForHabit(habit.getId());
+    public void markHabitAsCompleted(long habitId) throws SQLException {
+        List<LocalDate> completionHistory = habitComletionHistoryRepository.getCompletionHistoryForHabit(habitId);
+        HabitEntity habit = habitRepository.getHabitById(habitId);
 
-        if (completionHistory.isEmpty()
-                || !Objects.equals(completionHistory.get(completionHistory.size() - 1),
-                LocalDate.now())) {
-            habitComletionHistoryRepository.addComletionDateByHabitIdUserIs(
-                    habit.getId(), habit.getUserId(), LocalDate.now());
+        if (completionHistory.isEmpty() || !Objects.equals(completionHistory.get(completionHistory.size() - 1), LocalDate.now())) {
+            habitComletionHistoryRepository.addComletionDateByHabitIdUserIs(habit.getId(), habit.getUserId(), LocalDate.now());
         } else {
-            throw new IllegalArgumentException("Вы сегодня уже выполняли эту привычку");
+            throw new HabitAlreadyCompletedException("Вы сегодня уже выполняли эту привычку.");
         }
     }
 
@@ -126,7 +151,7 @@ public class HabitService {
      * @param period период ("day", "week", "month")
      * @return статистика выполнения привычки
      */
-    public int calculateHabitCompletedByPeriod(Habit habit, Period period) throws SQLException {
+    public int calculateHabitCompletedByPeriod(HabitEntity habit, Period period) throws SQLException {
         LocalDate now = LocalDate.now();
         LocalDate startDate = switch (period.getPeriodName().toLowerCase()) {
             case "day" -> now.minusDays(1);
@@ -150,7 +175,7 @@ public class HabitService {
      * @param habit привычка
      * @return текущий streak
      */
-    public int calculateCurrentStreak(Habit habit) throws SQLException {
+    public int calculateCurrentStreak(HabitEntity habit) throws SQLException {
         List<LocalDate> completions = habitComletionHistoryRepository.getCompletionHistoryForHabit(habit.getId())
                 .stream()
                 .sorted()
@@ -189,7 +214,7 @@ public class HabitService {
      * @param period период ("day", "week", "month")
      * @return процент успешного выполнения привычки
      */
-    public double calculateCompletionPercentage(Habit habit, Period period) throws SQLException {
+    public double calculateCompletionPercentage(HabitEntity habit, Period period) throws SQLException {
         LocalDate now = LocalDate.now();
         LocalDate startDate = switch (period.getPeriodName().toLowerCase()) {
             case "day" -> now.minusDays(1);
@@ -216,7 +241,7 @@ public class HabitService {
      * @param period период ("day", "week", "month")
      * @return отчет о прогрессе
      */
-    public HabitReportDto generateProgressReport(Habit habit, Period period) throws SQLException {
+    public HabitReportDto generateProgressReport(HabitEntity habit, Period period) throws SQLException {
         int streak = calculateCurrentStreak(habit);
         double completionPercentage = calculateCompletionPercentage(habit, period);
         int completionCount = calculateHabitCompletedByPeriod(habit, period);
